+++ src/entry.dev.tsx
@@ -11,7 +11,7 @@
  */
 import { type RenderOptions, render } from '@builder.io/qwik';
 
-import Root from './Root';
+import Root from './root';
 
 export default function (opts: RenderOptions) {
   return render(document, <Root />, opts);
