+++ src/entry.ssr.tsx
@@ -16,7 +16,7 @@ import {
 } from '@builder.io/qwik/server';
 import { manifest } from '@qwik-client-manifest';
 
-import Root from './Root';
+import Root from './root';
 
 export default function (opts: RenderToStreamOptions) {
   return renderToStream(<Root />, {
